+++ docs/.vitepress/theme/components/GlobalElement.vue
@@ -1,6 +1,6 @@
 <template>
   <FloatButton
-    :bottom="96"
+    :bottom="100"
     type="primary"
     tooltip="✨ 成为赞助者"
     :tooltip-props="{